+++ tasks/main.sh
@@ -21,7 +21,7 @@ main() {
     echo -e "${BOLD}${YELLOW}Prioritizing tasks with OpenAI for the next $days_to_process days...${RESET}"
     priorities=$(get_priorities "$day_tasks" "$days_to_process" "$start_day")
   else
-    echo -e "${BOLD}${YELLOW}Prioritizing tasks and events with OpenAI for today...${RESET}"
+    echo -e "${BOLD}${YELLOW}Prioritizing tasks and events with OpenAI for $start_day...${RESET}"
 
     priorities=$(get_priorities "$day_tasks" 1 "$start_day")
   fi
@@ -55,13 +55,15 @@ main() {
   # Add remaining hours
   remaining_hours=$(calculate_remaining_hours)
 
+  $file_path="$HOME/Documents/Brain dump/Päivän suunnittelu/$filename.md"
+
   # Save output to Obsidian vault with the current time and remaining hours in the header
-  echo -e "# $header\n\nKello on muistiinpanojen luomishetkellä $current_time. Päivää on jäljellä noin $remaining_hours tuntia. Yhteensä tapaamisia tänään $total_event_duration tuntia (mukaanlukien lounas). Päivässä aikaa tehtävien suorittamiseen jäljellä yhteensä $remaining_work_hours tuntia.\n\n## Päivän eventit$all_events\n\n$priorities" > "$HOME/Documents/Brain dump/Päivän suunnittelu/$filename.md"
+  echo -e "# $header\n\nKello on muistiinpanojen luomishetkellä $current_time. Päivää on jäljellä noin $remaining_hours tuntia. Yhteensä tapaamisia tänään $total_event_duration tuntia (mukaanlukien lounas). Päivässä aikaa tehtävien suorittamiseen jäljellä yhteensä $remaining_work_hours tuntia.\n\n## Päivän eventit$all_events\n\n$priorities" > $file_path
 
   # Add TASKS_TO_BE_SCHEDULED at the end of the file
-  echo -e "\n\n---\n\n## Aikataulutetut tehtävät\n\n\`\`\`\n$TASKS_TO_SCHEDULE\n\`\`\`" >> "$HOME/Documents/Brain dump/Päivän suunnittelu/$filename.md"
+  echo -e "\n\n---\n\n## Aikataulutetut tehtävät\n\n\`\`\`\n$TASKS_TO_SCHEDULE\n\`\`\`" >> $file_path
 
-  echo -e "${BOLD}${GREEN}Prioritization is ready and saved to Obsidian.${RESET}"
+  echo -e "${BOLD}${GREEN}Prioritization is ready and saved to Obsidian, file: $file_path.md${RESET}"
 
   # Debug: Print the full content of tasks to see what's being parsed
   if [ "$DEBUG" = true ]; then