+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove postpone function and combine with the schedule function
 * Add the list of tasks to be sheculed and links to tasks at the end of the note
 * Deprecate GENERAL_PROMPT
+* Do not sync focus events back to Todoist
 
 # 1.1.8: 2024-10-26